feat: make createPointOperation to be used in CypherFilter
